requirements.txt
streamlit
python-dotenv
requests
newspaper3k
requests-html
langchain
langchain-community
langchain-experimental
faiss-cpu
pandas
sentence-transformers
pypdf
duckduckgo-search
googlesearch-python
pydantic>=2,<3
beautifulsoup4

# Authentication and Security Dependencies
streamlit-authenticator==0.2.3
bcrypt==4.0.1
pyotp==2.9.0
qrcode==7.4.2.1
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
slowapi==0.1.9
redis==5.0.1
cryptography==41.0.7
